fix: clippy -D warnings across workspace
CI's `cargo clippy --workspace --all-targets -- -D warnings` is stricter
than the rustc warnings I ran locally. Five lints fired:

- geonative-geojson: IterInner::Streaming carried a 240 B Schema inline
  vs ~32 B for other variants — box it.
- geonative-geojson: outer `loop {}` in FeatureIter::next always returns;
  drop the unnecessary wrapper.
- geonative-pmtiles: redundant `'a` lifetime in find_tile — elide.
- geonative-geotiff: ok_or_else with a closure that's a plain constructor
  → use ok_or.
- geonative-geoparquet tests + extent.rs: `i as i32` where i is already
  i32 (from 0..N defaulting to i32) — drop the cast.

No behavioural changes; v0.4.0 release was rejected at validate stage
on these lints, no crates published yet.

+++ crates/geonative-geoparquet/src/extent.rs
@@ -145,7 +145,7 @@ mod tests {
                 let feat = Feature::new(
                     Some(i as i64 + 1),
                     Some(square_at(cx, cy, 0.1)),
-                    vec![Value::Int32(i as i32 + 1)],
+                    vec![Value::Int32(i + 1)],
                 );
                 w.write(&feat).unwrap();
             }
